fix: remove duplicate pom.xml
