+++ apps/wanaku-operator/src/main/java/ai/wanaku/operator/wanaku/WanakuTypes.java
@@ -1,8 +1,23 @@
 package ai.wanaku.operator.wanaku;
 
+import java.util.List;
+
 public final class WanakuTypes {
 
     private WanakuTypes() {}
+
+    // Deployment mode constants - must match CRD enum values
+    public static final String DEPLOYMENT_MODE_IN_CLUSTER = "in-cluster";
+    public static final String DEPLOYMENT_MODE_REMOTE = "remote";
+    public static final List<String> VALID_DEPLOYMENT_MODES =
+            List.of(DEPLOYMENT_MODE_IN_CLUSTER, DEPLOYMENT_MODE_REMOTE);
+
+    // Dependency cache strategy constants - must match CRD enum values
+    public static final String CACHE_STRATEGY_IN_MEMORY = "inmemory";
+    public static final String CACHE_STRATEGY_INFINISPAN = "infinispan";
+    public static final String CACHE_STRATEGY_DISABLED = "disabled";
+    public static final List<String> VALID_CACHE_STRATEGIES =
+            List.of(CACHE_STRATEGY_IN_MEMORY, CACHE_STRATEGY_INFINISPAN, CACHE_STRATEGY_DISABLED);
 
     public static class AuthSpec {
         private String authServer;
