+++ docs/testing-the-code.md
@@ -1,14 +1,40 @@
 # Wanaku Assertions Library
 
-Wanaku provides a domain-specific assertions helper designed to keep tests concise and to surface clear,
+Wanaku provides domain-specific assertions helpers designed to keep tests concise and to surface clear,
 actionable failure messages.
+
+## Core Assertions
 
 **Location**
 `core/core-util/src/test/java/ai/wanaku/test/assertions/WanakuAssertions.java`
 
 This class is shared across modules via the `core-util` test-jar (`classifier` `tests`).
 
+It provides assertions for:
+- HTTP responses (status, success/error)
+- Wanaku responses (success/error)
+- Tools, Resources, Forwards (registration, equality)
+- Containers (running, healthy)
+- Namespaces
+
+## Operator Assertions (Kubernetes-specific)
+
+**Location**
+`apps/wanaku-operator/src/test/java/ai/wanaku/operator/assertions/OperatorAssertions.java`
+
+This class provides Kubernetes-specific assertions for testing operator functionality:
+
+- `assertCondition(Condition, String, String, Long)` - Verify Kubernetes condition state
+- `assertServiceLabel(Service, String, String)` - Verify service has expected label
+- `assertServicePort(Service, int)` - Verify service exposes expected port
+- `assertMetadataLabel(HasMetadata, String, String)` - Verify resource has expected label
+- `assertEndpointTarget(Endpoints, String, int)` - Verify endpoint points to expected host/port
+
+These assertions are specifically for the operator module and use fabric8 Kubernetes client types.
+
 ## Usage
+
+### Core Assertions
 
 Add a static import in your test:
 
@@ -18,7 +44,20 @@ import static ai.wanaku.test.assertions.WanakuAssertions.assertNamespaceExists;
 import static ai.wanaku.test.assertions.WanakuAssertions.assertSuccessResponse;
 ```
 
+### Operator Assertions (Kubernetes-specific)
+
+For operator tests, use the operator-specific assertions:
+
+```java
+import static ai.wanaku.operator.assertions.OperatorAssertions.assertCondition;
+import static ai.wanaku.operator.assertions.OperatorAssertions.assertEndpointTarget;
+import static ai.wanaku.operator.assertions.OperatorAssertions.assertServiceLabel;
+import static ai.wanaku.operator.assertions.OperatorAssertions.assertServicePort;
+```
+
 ### Available helpers
+
+**Core Assertions (`WanakuAssertions`):**
 
 - `assertSuccessResponse(WanakuResponse<?>)`
 - `assertErrorResponse(WanakuResponse<?>)`
@@ -37,6 +76,14 @@ import static ai.wanaku.test.assertions.WanakuAssertions.assertSuccessResponse;
 - `assertContainerRunning(GenericContainer<?>)`
 - `assertContainerHealthy(GenericContainer<?>)`
 - `assertNamespaceExists(String, List<Namespace>)`
+
+**Operator Assertions (`OperatorAssertions`):**
+
+- `assertCondition(Condition, String, String, Long)` - Verify Kubernetes condition
+- `assertServiceLabel(Service, String, String)` - Verify service label
+- `assertServicePort(Service, int)` - Verify service port
+- `assertMetadataLabel(HasMetadata, String, String)` - Verify resource label
+- `assertEndpointTarget(Endpoints, String, int)` - Verify endpoint target
 
 ### Common patterns
